python_client/requirements.txt
customtkinter>=5.2.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
playwright>=1.40.0
playwright-stealth>=1.0.0
aiohttp>=3.9.0